+++ propaq/propagators/__init__.py
@@ -2,7 +2,8 @@
 
 from propaq._rust_core import PropagationResult as PropagationResult
 
-from ._abstract import AbstractPropagator as AbstractPropagator
+from .abstract import AbstractPropagator as AbstractPropagator
+from .abstract import CircuitLike as CircuitLike
 from .majorana import MajoranaPropagator as MajoranaPropagator
 from .pauli import PauliPropagator as PauliPropagator
 from .surrogate_majorana import MajoranaSurrogatePropagator as MajoranaSurrogatePropagator
@@ -15,6 +16,7 @@ __all__ = [
     "MajoranaPropagator",
     "PauliPropagator",
     "AbstractPropagator",
+    "CircuitLike",
     "PauliSurrogatePropagator",
     "MajoranaSurrogatePropagator",
     "PropagationResult",