+++ tests/propagator/test_pauli_loschmidt.py
@@ -5,7 +5,7 @@ import numpy as np
 from propaq.circuits import PauliCircuit
 from propaq.circuits.pauli.rotation import PauliRotation
 from propaq.datatypes import PauliString, PauliTermSum
-from propaq.datatypes._abstract import BitMask
+from propaq.datatypes.abstract import BitMask
 from propaq.propagators.pauli import PauliPropagator
 from propaq.truncation import TruncationPolicy